CitiBike/requirements.txt
flask==2.3.3
flask-cors==4.0.0
snowflake-connector-python==3.7.0
python-dotenv==1.0.0
google-generativeai==0.5.0